requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1654189308200/work
adabound==0.0.5
aiohttp @ file:///tmp/build/80754af9/aiohttp_1632748411861/work
albumentations==1.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
backcall==0.2.0
batchgenerators==0.25.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1607575130224/work
brotlipy==0.7.0
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1640686991047/work
captum==0.6.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1649184154184/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1651215153723/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1634230275410/work
cycler==0.10.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
efficientnet-pytorch==0.6.3
einops==0.6.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.17.0
flashtorch==0.1.3
fonttools==4.25.0
future==1.0.0
gdown==5.2.0
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1655273470589/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637590821884/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
ignite==1.1.0
imageio==2.13.5
imbalanced-learn==0.9.1
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1653252797772/work
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
linecache2==1.0.0
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1651821407140/work
MarkupSafe==2.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1634667019719/work
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
munch==2.5.0
munkres==1.1.4
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.6.3
nibabel==3.2.1
notebook==6.4.11
numexpr @ file:///tmp/build/80754af9/numexpr_1618856529730/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095651905/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
openpyxl @ file:///tmp/build/80754af9/openpyxl_1632777717936/work
orjson==3.7.2
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.13
pretrainedmodels==0.7.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.24.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydicom==2.3.0
Pygments==2.12.0
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1652398519695/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857263093/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
qudida==0.0.4
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1654796491502/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1637781155505/work
scikit-image==0.19.1
scikit-learn==1.1.3
scipy==1.7.3
seaborn @ file:///croot/seaborn_1673479180098/work
segmentation-models-pytorch==0.2.1
Send2Trash==1.8.0
SimpleITK==2.1.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
soupsieve==2.3.2.post1
stack-data==0.2.0
tensorboard @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorboard_1678126660189/work/tensorboard-2.11.0-py3-none-any.whl
tensorboard-data-server @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-data-server_1622834972091/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX==2.6.2.2
terminado==0.13.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tinycss2==1.1.1
torch==1.10.0
torchaudio==0.10.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.11.1
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traceback2==1.4.0
traitlets==5.1.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
unittest2==1.1.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
widgetsnbextension==3.6.0
wrapt==1.14.1
yarg==0.1.9
yarl @ file:///tmp/build/80754af9/yarl_1606939947528/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
